refactor tests into individual files
